Add multi-select support.
This only works in Firefox, since Chrome currently doesn't support this
API. As a result, when a user opens a link that has multiple selections,
I currently just alert() a error message explaining this (and highlight
just the first one).

This isn't a great solution, but I don't think a great solution exists
here. Here are some alternatives:

* Disallow multiselect. If you do this you have to make a choice - do
  you want to make it so the user can use multiselect, and just don't
  generate a hash when they do? Do you generate a hash, but just to one
  of the selections? If so, which selection? How do you communicate that
  to the user, if at all? There's also the option of preventing the user
  from using multiselect at all on the page, but that's a bad option.
  Don't do that.
* Fall back to highlighting by inserting a styled <mark> tag or similar.
  This is too complicated for my tastes.
* Display the entire selected text in the alert box, but only actually
  select one of the highlights.
* Fail silently. Either highlight only one of them, or don't highlight
  any at all.

In the future, I may pick a few of these strategies that I think are not
so fundamentally bad that they shouldn't exist, and allow configuration
over which strategy is chosen.

I may also consider this to be a obscure enough edge case to go move the
error handling logic to a separate, dynamically loaded module, so that
the main bundle doesn't have to include it.