Change strategy for handling duplicate nodes.
Previously, only offsets were used, but it's challenging to handle
changing documents strictly via offsets — it's not possible in the
general case to detect if a document has changed, let alone handle it in
a reasonable way. The new approach is encode a list of all duplicate
nodes in the URL, which contains strictly more information, which should
make it easier to handle edge cases in the future.