Enable selecting non-text nodes.
This adds significant complexity to the code, but it's needed to make
things like triple-click select work.

In the future, it's possible that a better approach will be to enable
hashing of non-text nodes, but that's unlikely to be the best solution
for a few reasons:

* Determining how to hash such nodes is really hard — for a `<img>` node,
  the `src` attribute should probably matter, but should the `class` or
  `style` attribute? For that matter, if the `src` changes, but the
  contents of the image is the same, shouldn't that be the same image?
  These sorts of questions are repeated for nearly every type of node.
* It's much more likely that there will be duplicate hashes this way,
  which is not a great situation.

Doing it this way has downsides — complexity, the fact that in some
browsers selecting block elements is distinct from selecting text, etc,
but for now, the tradeoffs seem worth it.